Еще один шаг на пути к абстракции, компилируемый, но не рабочий
