+++ OnlineSchoolSystem.Client/OnlineSchoolSystem.Client.csproj
@@ -6,7 +6,10 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\OnlineSchoolSystem.DataAccess.File\OnlineSchoolSystem.DataAccess.FileStorage.csproj" />
+    <ProjectReference Include="..\OnlineSchoolSystem.Models\OnlineSchoolSystem.Models.csproj" />
     <ProjectReference Include="..\OnlineSchoolSystem.Utilites\OnlineSchoolSystem.Utilites.csproj" />
+    <ProjectReference Include="..\OnlineSchoolSystem.YoutubeBot\OnlineSchoolSystem.Bots.csproj" />
   </ItemGroup>
