+++ OnlineSchoolSystem.DataAccess.File/OnlineSchoolSystem.DataAccess.FileStorage.csproj
@@ -10,7 +10,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\OnlineSchoolSystem.Models\OnlineSchoolSystem.Models.csproj" />
-    <ProjectReference Include="..\OnlineSchoolSystem.YoutubeBot\OnlineSchoolSystem.Bots.csproj" />
   </ItemGroup>
 
 </Project>
