--- OnlineSchoolSystem.Models/IMember.cs
@@ -1,6 +1,0 @@
-﻿namespace OnlineSchoolSystem.Models
-{
-    public interface IMember
-    {
-    }
-}
