+++ OnlineSchoolSystem.Models/IStorage.cs
@@ -1,6 +1,9 @@
-﻿namespace OnlineSchoolSystem.Models
+﻿using System.Collections.Generic;
+
+namespace OnlineSchoolSystem.Models
 {
     public interface IStorage
     {
+        void Store(IEnumerable<Message> messages);
     }
 }
