+++ OnlineSchoolSystem.Models/Member.cs
@@ -1,0 +1,7 @@
+﻿namespace OnlineSchoolSystem.Models
+{
+    public class Member
+    {
+        public string Name { get; set; }
+    }
+}
